Move shipping mutations to a separate file + cleanup
