+++ src/v2/Apps/Order/Components/AddressModal.tsx
@@ -1,10 +1,6 @@
 import React from "react"
 import { Button, Input, Modal, Spacer } from "@artsy/palette"
-import {
-  convertShippingAddressForExchange,
-  convertShippingAddressToMutationInput,
-  SavedAddressType,
-} from "../Utils/shippingAddressUtils"
+import { SavedAddressType } from "../Utils/shippingAddressUtils"
 import { FormikProps, useFormik } from "formik"
 import {
   removeEmptyKeys,
@@ -12,9 +8,8 @@ import {
   validatePhoneNumber,
 } from "../Utils/formValidators"
 import { CountrySelect } from "v2/Components/CountrySelect"
-import { graphql } from "react-relay"
-import { AddressModalMutation } from "v2/__generated__/AddressModalMutation.graphql"
 import { CommitMutation } from "../Utils/commitMutation"
+import { updateUserAddress } from "../Mutations/shippingMutations"
 
 interface Props {
   show: boolean
@@ -23,63 +18,6 @@ interface Props {
   commitMutation: CommitMutation
   onSuccess: (address) => void
   onError: (message: string) => void
-}
-
-const saveAddress = async (
-  commitMutation: CommitMutation,
-  userAddressID: string,
-  values: any,
-  closeModal: () => void,
-  onSuccess: (address) => void,
-  onError: (message: string) => void
-) => {
-  const useArtrubutes = convertShippingAddressToMutationInput(values)
-
-  const result = await commitMutation<AddressModalMutation>({
-    variables: {
-      input: {
-        userAddressID: userAddressID,
-        attributes: useArtrubutes,
-      },
-    },
-    mutation: graphql`
-      mutation AddressModalMutation($input: UpdateUserAddressInput!) {
-        updateUserAddress(input: $input) {
-          userAddressOrErrors {
-            ... on UserAddress {
-              id
-              internalID
-              name
-              addressLine1
-              addressLine2
-              isDefault
-              phoneNumber
-              city
-              region
-              postalCode
-              country
-            }
-            ... on Errors {
-              errors {
-                code
-                message
-              }
-            }
-          }
-        }
-      }
-    `,
-  })
-  const errors = result.updateUserAddress.userAddressOrErrors.errors
-  if (errors) {
-    onError(errors.map(error => error.message).join(", "))
-  } else {
-    const address = convertShippingAddressForExchange(
-      result.updateUserAddress.userAddressOrErrors
-    )
-    onSuccess(address)
-  }
-  closeModal()
 }
 
 const validateor = (values: any) => {
@@ -104,7 +42,7 @@ export const AddressModal: React.FC<Props> = ({
     initialValues: address,
     validate: validateor,
     onSubmit: values => {
-      saveAddress(
+      updateUserAddress(
         commitMutation,
         address.internalID,
         values,
